+++ R/on-load.R
@@ -11,6 +11,7 @@
   })
 }
 
+
 # Evaluate an expression, and if it emits any messsages, convert them to
 # packageStartupMessage.
 convert_message_to_package_startup_message <- function(expr) {
@@ -18,7 +19,16 @@ convert_message_to_package_startup_message <- function(expr) {
     force(expr),
     message = function(cnd) {
       packageStartupMessage(conditionMessage(cnd))
-      tryInvokeRestart("muffleMessage")
+      maybe_restart("muffleMessage")
     }
   )
+}
+
+
+# This is equivalent to tryInvokeRestart, but that function is available only in
+# R 4.0 and above.
+maybe_restart <- function(restart) {
+  if (!is.null(findRestart(restart))) {
+    invokeRestart(restart)
+  }
 }